+++ src/energy_systems/storage/battery.jl
@@ -16,6 +16,7 @@ Base.@kwdef mutable struct Battery <: Component
 
     capacity::Float64
     load::Float64
+    load_end_of_last_timestep::Float64
     losses::Float64
 
     max_charge::Float64
@@ -33,6 +34,7 @@ Base.@kwdef mutable struct Battery <: Component
                    medium,
                    config["capacity"], # capacity
                    config["load"],     # load
+                   0.0, # load_end_of_last_timestep
                    0.0, # losses
                    0.0, # max_charge
                    0.0) # max_discharge
@@ -44,6 +46,8 @@ function initialise!(unit::Battery, sim_params::Dict{String,Any})
                           unload_storages(unit.controller, unit.medium))
     set_storage_transfer!(unit.output_interfaces[unit.medium],
                           load_storages(unit.controller, unit.medium))
+
+    unit.load_end_of_last_timestep = copy(unit.load)
 end
 
 function reset(unit::Battery)
@@ -76,11 +80,12 @@ end
 function balance_on(interface::SystemInterface,
                     unit::Battery)::Vector{EnergyExchange}
     caller_is_input = unit.uac == interface.target.uac
+    balance_written = interface.max_energy.max_energy[1] === nothing || interface.sum_abs_change > 0.0
     purpose_uac = unit.uac == interface.target.uac ? interface.target.uac : interface.source.uac
 
     return [EnEx(;
                  balance=interface.balance,
-                 energy_potential=caller_is_input ? -(unit.max_charge) : unit.max_discharge,
+                 energy_potential=balance_written ? 0.0 : (caller_is_input ? -(unit.max_charge) : unit.max_discharge),
                  purpose_uac=purpose_uac,
                  temperature_min=interface.temperature_min,
                  temperature_max=interface.temperature_max,
@@ -115,6 +120,7 @@ end
 function load(unit::Battery, sim_params::Dict{String,Any})
     if unit.max_charge < sim_params["epsilon"]
         set_max_energy!(unit.output_interfaces[unit.medium], 0.0)
+        unit.load_end_of_last_timestep = copy(unit.load)
         return
     end
 
@@ -124,6 +130,7 @@ function load(unit::Battery, sim_params::Dict{String,Any})
 
     if energy_available <= 0.0
         set_max_energy!(unit.input_interfaces[unit.medium], 0.0)
+        unit.load_end_of_last_timestep = copy(unit.load)
         return # load is only concerned with receiving energy from the source
     end
 
@@ -135,6 +142,8 @@ function load(unit::Battery, sim_params::Dict{String,Any})
         unit.load = unit.capacity
         sub!(inface, diff)
     end
+
+    unit.load_end_of_last_timestep = copy(unit.load)
 end
 
 function output_values(unit::Battery)::Vector{String}